Handle the issue in which the UI gets reset after an upgrade of the base app
The current version of the plugin assumes that when you upgrade the base app,
you also upgrade the UI. That is a fine assumption for service-focused apps,
which presumably want to expose the new functionality in a user-visible
fashion, but it is not a good fit for us because:
- we may want to tweak the data collection regime
- given that people deploy their own copies of the UI, we may not be able to
  coordinate with all of them to update all the UIs

We really want to keep the UI and update the underlying app. Of course, this
assumes that there are no API changes.

Fortunately, the change was pretty straightforward.